+++ src/index.ts
@@ -1,6 +1,9 @@
 import { setCanvasWindowSize } from './canvas'
 import { renderConfetti } from './confetti'
+
 export { createConfetti as create } from './confetti'
 
-window.addEventListener('resize', () => { setCanvasWindowSize() })
-window.requestAnimationFrame(renderConfetti)
+if(window) {
+    window.addEventListener('resize', () => { setCanvasWindowSize() })
+    window.requestAnimationFrame(renderConfetti)
+}