+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Changelog
 
+### 0.2.0
+
+Release 2 oct 2023
+
+  * Added option to disable ticks (opacity). Put -1 on ticks.
+  * Changed the z-index by default of the canvas. Now is  Number.MAX_SAFE_INTEGER (2^53 − 1).
+  * Added window checks to avoid server execution errors (SSR).
+
 ### 0.1.2
 
 Release 2 oct 2023
+
   * Added particle position check to delete it from array when it comes out of canvas.
 
 ### 0.1.0
 
 Release 1 oct 2023
+
   * First version